requirements.txt
django-cors-headers==4.1.0
environs==9.5.0
django-cleanup==8.0.0
drf_yasg==1.21.7
gunicorn==21.2.0
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.1.0
django-jazzmin==2.6.0